requirements.txt
fastapi
uvicorn
streamlit
langgraph
langchain
google-api-python-client
google-auth
google-auth-oauthlib
requests
